Rename file to match exposed function name
